Revamp files import

+++ 7.1/actions/init-wordpress.sh
@@ -18,16 +18,20 @@ ${disclaimer} \\
 require_once '${WODBY_DIR_CONF}/wodby.wp-config.php';" "${wp_config}"
 fi
 
-# Sync and symlink files dir if it's not symlink already.
+# Symlink files dir
 WP_FILES="${WP_ROOT}/wp-content/uploads"
-mkdir -p "${WP_FILES}"
 
-if [[ ! -L "${WP_FILES}" ]]; then
-    mkdir -p "${WODBY_DIR_FILES}/private"
-    mkdir -p "${WODBY_DIR_FILES}/public"
-
-    rsync -rlt --force "${WP_FILES}" "${WODBY_DIR_FILES}/public/"
-
-    rm -rf "${WP_FILES}"
+if [[ -d "${WP_FILES}" ]]; then
+    if [[ ! -L "${WP_FILES}" ]]; then
+        if [[ "$(ls -A "${WP_FILES}")" ]]; then
+            echo "Error: directory ${WP_FILES} exists and is not empty. The files directory can not be under version control or must be empty."
+            exit 1
+        # If dir is not symlink and empty, remove it and link.
+        else
+            rm -rf "${WP_FILES}"
+            ln -sf "${WODBY_DIR_FILES}/public" "${WP_FILES}"
+        fi
+    fi
+else
     ln -sf "${WODBY_DIR_FILES}/public" "${WP_FILES}"
 fi
